Move my_node to package
