+++ src/p2p/src/my_node/__init__.py
@@ -1,0 +1,6 @@
+# src/p2p/src/my_node/__init__.py
+# For importing purposes
+
+
+# Imports
+from .node_list import node_list
